* Lots of files: Getting rid of warning due to type conversion, unused 	  args and other such stuff.
	* convert/u8/ConvertUsecodeU8.h : Marking processes, and fixing a few 
	  miskates.
	  
	* graphics/ShapeInfo.h, graphics/TypeFlags.cpp : Added some new unknown
	  typeflags.
	  
	* kernel/Application.cpp, misc/Console.* : Clicking an item will return its
	  quality. Added ability to scroll console with Pageup and Pagedown.

	* world/Item.cpp : Changed the NPC classid detecion in callUsecodeEvent 
	  since npcs aren't the only items to have somehing in npcnum. Fixed up
	  I_getTypeFlag which was seriously broken. The new code is horrible
	  to look at.
	  
	* world/Item.h : Added FLG_MONSTER_NPC. But, I'm not sure if that is 
	  correct.
	  
	* world/ItemSorter.cpp : Added a few consts. A few other minor changes.
	
	* usecode/u8intrinsics.h : All unimplemented processes should now be dummy 
	  processes.

+++ usecode/UCList.cpp
@@ -49,8 +49,8 @@ void UCList::copyStringList(UCList& l)
 	for (unsigned int i = 0; i < l.size; i++) {
 		uint16 s = ucm->duplicateString(l.getStringIndex(i));
 		uint8 tmp[2]; // ugly...
-		tmp[0] = s & 0xFF;
-		tmp[1] = s >> 8;
+		tmp[0] = static_cast<uint8>(s & 0xFF);
+		tmp[1] = static_cast<uint8>(s >> 8);
 		append(tmp);
 	}
 }
@@ -93,8 +93,8 @@ void UCList::assignString(uint32 index, uint16 str)
 	// free old contents of element index; take ownership of str(?)
 
 	UCMachine::get_instance()->freeString(getStringIndex(index));
-	elements[index*elementsize] = str & 0xFF;
-	elements[index*elementsize+1] = str >> 8;
+	elements[index*elementsize] = static_cast<uint8>(str & 0xFF);
+	elements[index*elementsize+1] = static_cast<uint8>(str >> 8);
 }
 
 void UCList::removeString(uint16 s)
